Rename some files bucket -> bins
