+++ benchmark/transform_benchmark.cpp
@@ -6,8 +6,7 @@
 
 #include <random>
 
-#include "scipp/variable/bucket_model.h"
-#include "scipp/variable/buckets.h"
+#include "scipp/variable/bins.h"
 #include "scipp/variable/transform.h"
 #include "scipp/variable/variable.h"
 
@@ -215,8 +214,7 @@ static void BM_transform_buckets_inplace_unary(benchmark::State &state) {
     n_element += size;
   }
   Variable buffer = makeVariable<double>(Dims{Dim::X}, Shape{n_element});
-  Variable var{std::make_unique<variable::DataModel<bucket<Variable>>>(
-      indices, Dim::X, buffer)};
+  Variable var = make_bins(indices, Dim::X, buffer);
 
   for (auto _ : state) {
     transform_in_place<double>(var, [](auto &x) { x += x; });
