+++ dataset/CMakeLists.txt
@@ -2,8 +2,8 @@
 # contributors (https://github.com/scipp)
 set(TARGET_NAME "scipp-dataset")
 set(INC_FILES
+    include/scipp/dataset/bins.h
     include/scipp/dataset/bucketby.h
-    include/scipp/dataset/bucket.h
     include/scipp/dataset/choose.h
     include/scipp/dataset/counts.h
     include/scipp/dataset/dataset_access.h
@@ -26,8 +26,8 @@ set(INC_FILES
 
 set(SRC_FILES
     arithmetic.cpp
+    bins.cpp
     bucketby.cpp
-    bucket.cpp
     counts.cpp
     data_array.cpp
     dataset_access.cpp
