+++ dataset/bins.cpp
@@ -10,15 +10,10 @@
 #include "scipp/core/element/histogram.h"
 #include "scipp/core/except.h"
 #include "scipp/core/histogram.h"
-#include "scipp/dataset/bins_view.h"
-#include "scipp/dataset/bucket.h"
-#include "scipp/dataset/dataset.h"
-#include "scipp/dataset/histogram.h"
-#include "scipp/dataset/reduction.h"
-#include "scipp/dataset/shape.h"
+
 #include "scipp/variable/arithmetic.h"
+#include "scipp/variable/bins.h"
 #include "scipp/variable/bucket_model.h"
-#include "scipp/variable/buckets.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/shape.h"
 #include "scipp/variable/subspan_view.h"
@@ -26,6 +21,13 @@
 #include "scipp/variable/transform_subspan.h"
 #include "scipp/variable/util.h"
 #include "scipp/variable/variable_factory.h"
+
+#include "scipp/dataset/bins.h"
+#include "scipp/dataset/bins_view.h"
+#include "scipp/dataset/dataset.h"
+#include "scipp/dataset/histogram.h"
+#include "scipp/dataset/reduction.h"
+#include "scipp/dataset/shape.h"
 
 #include "../variable/operations_common.h"
 #include "dataset_operations_common.h"
