+++ dataset/include/scipp/dataset/bins.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "scipp/dataset/dataset.h"
-#include "scipp/variable/buckets.h"
+#include "scipp/variable/bins.h"
 
 namespace scipp::dataset {
