+++ dataset/test/CMakeLists.txt
@@ -5,8 +5,8 @@ add_dependencies(all-tests ${TARGET_NAME})
 add_executable(
   ${TARGET_NAME} EXCLUDE_FROM_ALL
   attributes_test.cpp
+  bins_test.cpp
   bucketby_test.cpp
-  buckets_test.cpp
   choose_test.cpp
   concatenate_test.cpp
   coords_view_test.cpp
