+++ dataset/test/size_of_test.cpp
@@ -2,11 +2,11 @@
 // Copyright (c) 2020 Scipp contributors (https://github.com/scipp)
 #include <gtest/gtest.h>
 
-#include "scipp/dataset/bucket.h"
+#include "scipp/dataset/bins.h"
 #include "scipp/dataset/dataset.h"
 #include "scipp/dataset/util.h"
 #include "scipp/variable/arithmetic.h"
-#include "scipp/variable/buckets.h"
+#include "scipp/variable/bins.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/util.h"
