+++ neutron/convert.cpp
@@ -10,7 +10,7 @@
 #include "scipp/variable/transform.h"
 #include "scipp/variable/util.h"
 
-#include "scipp/dataset/bucket.h"
+#include "scipp/dataset/bins.h"
 #include "scipp/dataset/dataset.h"
 #include "scipp/dataset/dataset_util.h"
