+++ variable/include/scipp/variable/bins.h
