+++ variable/test/bucket_model_test.cpp
@@ -3,8 +3,8 @@
 #include <gtest/gtest.h>
 
 #include "scipp/variable/arithmetic.h"
+#include "scipp/variable/bins.h"
 #include "scipp/variable/bucket_model.h"
-#include "scipp/variable/buckets.h"
 
 using namespace scipp;
 using namespace scipp::variable;
