+++ variable/test/variable_test.cpp
@@ -6,7 +6,7 @@
 #include "test_macros.h"
 
 #include "scipp/core/except.h"
-#include "scipp/variable/buckets.h"
+#include "scipp/variable/bins.h"
 #include "scipp/variable/operations.h"
 #include "scipp/variable/shape.h"
 #include "scipp/variable/variable.h"
